Setting the double as long and implementing optional manual commit

+++ src/main/java/com/revature/CLIBank/Repository/AccountRepo.java
@@ -43,7 +43,7 @@ public class AccountRepo {
                     userID TEXT NOT NULL,
                     pin INTEGER NOT NULL,
                     accountType TEXT NOT NULL,
-                    balance REAL NOT NULL DEFAULT 0.0,
+                    balance INTEGER NOT NULL DEFAULT 0,
                     frozen INTEGER NOT NULL DEFAULT 0,
                     FOREIGN KEY (userID) REFERENCES Owners(userID)
                 );
@@ -71,7 +71,7 @@ public class AccountRepo {
                 sourceAccountId TEXT NOT NULL,
                 destinationAccountId TEXT,
                 type TEXT NOT NULL,
-                amount REAL NOT NULL,
+                amount INTEGER NOT NULL,
                 timestamp TEXT NOT NULL,
                 FOREIGN KEY (sourceAccountId) REFERENCES Accounts(accountID),
                 FOREIGN KEY (destinationAccountId) REFERENCES Accounts(accountID)
@@ -138,7 +138,7 @@ public class AccountRepo {
                 ps.setString(2, account.getUserID().toString());
                 ps.setInt(3, account.getPin());
                 ps.setString(4, account.getAccountType().name());
-                ps.setDouble(5, account.getBalance());
+                ps.setLong(5, account.getBalance());
                 ps.setInt(6, account.isFrozen()? 1 : 0);
                 ps.executeUpdate();
     
@@ -172,7 +172,7 @@ public class AccountRepo {
             }
 
             ps.setString(4, transaction.getType().name());
-            ps.setDouble(5, transaction.getAmount());
+            ps.setLong(5, transaction.getAmount());
             ps.setString(6, transaction.getTimestamp().toString());
             ps.executeUpdate();
 
@@ -205,7 +205,7 @@ public class AccountRepo {
             ownerId == null ? null : UUID.fromString(ownerId),
             rs.getInt("pin"),
             AccountType.valueOf(rs.getString("accountType")),
-            rs.getDouble("balance"),
+            rs.getLong("balance"),
             rs.getInt("frozen") != 0
         );
     }
@@ -220,7 +220,7 @@ public class AccountRepo {
             UUID.fromString(rs.getString("sourceAccountId")),
             rs.getString("destinationAccountId") == null ? null : UUID.fromString(rs.getString("destinationAccountId")),
             TransactionType.valueOf(rs.getString("type")),
-            rs.getDouble("amount"),
+            rs.getLong("amount"),
             LocalDateTime.parse(rs.getString("timestamp"))
         );
     }
@@ -374,7 +374,7 @@ public class AccountRepo {
             ps.setString(1, account.getUserID().toString());
             ps.setInt(2, account.getPin());
             ps.setString(3, account.getAccountType().name());
-            ps.setDouble(4, account.getBalance());
+            ps.setLong(4, account.getBalance());
             ps.setInt(5, account.isFrozen() ? 1 : 0);
             ps.setString(6, account.getAccountID().toString());
             ps.executeUpdate();
@@ -383,6 +383,39 @@ public class AccountRepo {
             e.printStackTrace();
         }
     }
+
+
+    /**
+     * Method to take money from one account and insert it into another account
+     */
+    // public static void transferMoney(AccountInfo source, AccountInfo dest, long amount) {
+    //     String addSQL = "UPDATE Accounts SET balance = balance + ? WHERE accountID = ?";
+    //     String withdrawSQL = "UPDATE Accounts SET balance = balance - ? WHERE accountID = ?";
+
+    //     try(Connection connection = ConnectionFactory.getManualCommitConnection()) {
+    //         try(PreparedStatement ps = connection.prepareStatement(addSQL)) {
+    //             ps.setLong(1, amount);
+    //             ps.setString(2, dest.getAccountID().toString());
+    //             int rowCount = ps.executeUpdate();
+    //             if (rowCount != 1){
+    //                 connection.rollback();
+    //                 throw new SQLException("Adding money to account failed");
+    //             }
+    //         }
+    //         try (PreparedStatement ps2 = connection.prepareStatement(withdrawSQL)) {
+    //             ps2.setLong(1, amount);
+    //             ps2.setString(2, source.getAccountID().toString());
+    //             int rowCount = ps2.executeUpdate();
+    //             if(rowCount != 1) {
+    //                 connection.rollback();
+    //                 throw new SQLException("Withdrawing money from account failed");
+    //             }
+    //         }
+    //         connection.commit();
+    //     } catch (SQLException e) {
+    //         e.printStackTrace();
+    //     }
+    // }
